requirements.txt
anyio==4.9.0
base58==2.1.1
certifi==2025.1.31
Deprecated==1.2.18
dotenv==0.9.9
ECPy==1.2.5
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pycryptodome==3.22.0
python-dotenv==1.0.1
sniffio==1.3.1
types-Deprecated==1.2.15.20250304
typing_extensions==4.12.2
websockets==15.0.1
wrapt==1.17.2
xrpl-py==4.1.0
